feat (CommandLine): Allow option to set baseappdatapath command param  #779 (#877)
* feat (BaseAppPath): Allow option to set baseAppData path which allows customisation of config and db relative path.  #779

baseappdata - Set the base app data folder, which will be use to read config and relative db location.
NOTE: (basepath) is already used to set virtual path for webroot.
Removed code in startup that was migrating the db to new location.  Since dbname can be configured this code is possibly confusing and superceded.

* fix (CommandLineOptions): rename to 'baseappdatapath' to be consistent with other CommandLineOptions.

* Clarify appdata location in settings docs.

* Fix bad rebase.

Co-authored-by: Rob Wood <rob@rnwood.co.uk>