feature(relay): Implement Javascript expression to make auto relay rules more flexible. (#1347)
This allows you to write an expression to control if a message is relayed or not based on its metadata or info from the session.

It also allows the recipients to be dynamically modified when relaying. For example replacing @real.com with @test.com.

See the comments in appsettings.json for full details on how to use this.

This has also been exposed in the settings dialog.